+++ serde-encrypt/tests/example_serde_encrypt_public_key_struct_with_reference.rs
@@ -4,10 +4,10 @@ use serde::{Deserialize, Serialize};
 use serde_encrypt::{
     serialize::{impls::CborSerializer, TypedSerialized},
     traits::SerdeEncryptPublicKey,
+    Error,
 };
 use serde_encrypt_core::{
     encrypt::encrypted_message::EncryptedMessage,
-    error::Error,
     key::{
         combined_key::{ReceiverCombinedKey, SenderCombinedKey},
         key_pair::{ReceiverKeyPair, SenderKeyPair},