+++ serde-encrypt/tests/example_serde_encrypt_shared_key_encryption_with_key_exchange.rs
@@ -19,10 +19,10 @@ use serde_encrypt::{
     serialize::{impls::CborSerializer, TypedSerialized},
     shared_key::SharedKey,
     traits::{SerdeEncryptPublicKey, SerdeEncryptSharedKey},
+    Error,
 };
 use serde_encrypt_core::{
     encrypt::encrypted_message::EncryptedMessage,
-    error::Error,
     key::{
         as_shared_key::AsSharedKey,
         combined_key::{ReceiverCombinedKey, SenderCombinedKey},