+++ serde-encrypt/tests/example_serde_encrypt_shared_key_owned_data.rs
@@ -2,10 +2,10 @@
 
 use serde::{Deserialize, Serialize};
 use serde_encrypt::{
-    serialize::impls::CborSerializer, shared_key::SharedKey, traits::SerdeEncryptSharedKey,
+    serialize::impls::CborSerializer, shared_key::SharedKey, traits::SerdeEncryptSharedKey, Error,
 };
 use serde_encrypt_core::{
-    encrypt::encrypted_message::EncryptedMessage, error::Error, key::as_shared_key::AsSharedKey,
+    encrypt::encrypted_message::EncryptedMessage, key::as_shared_key::AsSharedKey,
 };
 
 #[derive(Debug, Serialize, Deserialize)]