+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 
 ## [Unreleased]
 
+### Changed
+
+- `serde_error::error::Error` to `serde_error::Error`.
+
 ---
 
 ## [Released]